Renamed pkg test to testutils.
